修复部分缺陷

+++ src/components/project/PersonnelList.vue
@@ -1,66 +1,52 @@
 <template>
-    <div>
-        <div class="card">
-            <div class="card-close">
-                <div class="dropdown">
-                    <button aria-expanded="false" aria-haspopup="true" class="dropdown-toggle" data-toggle="dropdown"
-                            id="closeCard3" type="button"><i class="fa fa-ellipsis-v"></i></button>
-                    <div aria-labelledby="closeCard3" class="dropdown-menu dropdown-menu-right has-shadow"><a
-                        class="dropdown-item remove"
-                        href="#">
-                        <i class="fa fa-times"></i>Close</a><a class="dropdown-item edit" href="#"> <i
-                        class="fa fa-gear"></i>Edit</a></div>
-                </div>
-            </div>
-            <div class="card-header d-flex align-items-center">
-                <h3 class="h4">项目人员</h3>
-                <button @click="addPersonnel">添加</button>
-            </div>
-            <div class="card-body">
-                <div class="table-responsive">
-                    <table class="table table-striped table-hover">
-                        <thead>
-                        <tr>
-                            <th>序号</th>
-                            <th>姓名</th>
-                            <th>职位</th>
-                        </tr>
-                        </thead>
-                        <tbody>
-                        <tr :key="index" v-for="(personnel,index) in personnelList">
-                            <th scope="row">{{index + 1}}</th>
-                            <td>{{personnel.userId}}</td>
-                            <td>{{getRoleTypeName(personnel.roleType)}}</td>
-                            <button @click="edit(personnel)">编辑</button>
-                            <button @click="remove(personnel.id)">删除</button>
-                        </tr>
-                        </tbody>
-                    </table>
-                </div>
-            </div>
-        </div>
+    <table class="table table-striped table-hover">
+        <thead>
+        <tr>
+            <th>序号</th>
+            <th>姓名</th>
+            <th>职位</th>
+        </tr>
+        </thead>
         <el-dialog
             :title="dialogTitle"
-            :visible.sync="dialogVisible"
+            :visible="dialogVisible"
             @close="dialogVisible = false">
             <personnel-editor
+                :nickname="dialogNickname"
                 :personnel="dialogPersonnel"
                 @save-personnel="onSavePersonnel">
             </personnel-editor>
         </el-dialog>
-    </div>
+        <tbody>
+        <tr :key="index" v-for="(personnel,index) in personnelList">
+            <th scope="row">{{index + 1}}</th>
+            <td>{{getUsername(personnel.userId)}}</td>
+            <td>{{getRoleTypeName(personnel.roleType)}}</td>
+            <td>
+                <button @click="edit(personnel)" class="btn btn-primary btn-sm">编辑</button>
+                <button @click="remove(personnel.id)" class="btn btn-primary btn-sm">删除</button>
+            </td>
+        </tr>
+        </tbody>
+    </table>
+
 </template>
 
 <script>
     import PersonnelService from '../../service/PersonnelService'
     import RoleTypeNameMap from '../../entity/RoleTypeNameMap'
     import PersonnelEditor from './PersonnelEditor'
+    import EventBus from '../../util/EventBus'
+    import UserService from '../../service/UserService'
 
     export default {
         name: 'PersonnelList',
         components: {PersonnelEditor},
         created () {
-            this.getPersonnelList()
+            this.getData()
+            EventBus.$on('refresh-personnel-list', () => {
+                this.getData()
+            })
         },
         props: {
             projectId: ''
@@ -71,56 +57,64 @@
                 personnelList: [],
                 dialogTitle: '',
                 dialogVisible: false,
-                dialogPersonnel: ''
+                dialogPersonnel: '',
+                dialogNickname: '',
+                userInfoList: []
             }
         },
         watch: {
             projectId () {
-                this.getPersonnelList()
-            },
-            dialogVisible () {
-                this.getPersonnelList()
+                this.getData()
             }
         },
         methods: {
             getPersonnelList () {
-                if (this.projectId === null || this.projectId === '') {
-                    return
-                }
                 PersonnelService.getPersonnelList(this.projectId, data => {
                     this.personnelList = data
                 })
             },
+            getUserInfoList (callback) {
+                let self = this
+                UserService.getUserInfoOfProject(this.projectId, data => {
+                    self.userInfoList = data
+                    callback()
+                })
+            },
+            getData () {
+                this.getUserInfoList(() => {
+                    this.getPersonnelList()
+                })
+            },
+            getUsername (userId) {
+                for (let i = 0; i < this.userInfoList.length; i++) {
+                    if (this.userInfoList[i].id === userId) {
+                        return this.userInfoList[i].nickname
+                    }
+                }
+                return null
+            },
             getRoleTypeName (roleType) {
                 return this.map.get(roleType)
             },
-            addPersonnel () {
-                this.dialogTitle = '新增'
-                this.dialogPersonnel = {
-                    projectId: this.projectId
-                }
-                this.dialogVisible = true
-            },
             edit (personnel) {
                 this.dialogTitle = '编辑'
+                this.dialogNickname = this.getUsername(personnel.userId)
                 this.dialogPersonnel = personnel
                 this.dialogVisible = true
             },
-            onSavePersonnel (personnel) {
-                if (this.dialogTitle === '新增') {
-                    PersonnelService.addPersonnel(personnel, () => {
-                        this.dialogVisible = false
-                    })
-                } else if (this.dialogTitle === '编辑') {
-                    PersonnelService.updatePersonnel(personnel, () => {
-                        this.dialogVisible = false
-                    })
-                }
-            },
             remove (personnelId) {
+                if (!confirm('是否删除？')) {
+                    return
+                }
                 PersonnelService.removePersonnel(personnelId, () => {
                     this.dialogVisible = false
-                    this.getPersonnelList()
+                    this.getData()
+                })
+            },
+            onSavePersonnel (personnel) {
+                PersonnelService.updatePersonnel(personnel, () => {
+                    this.getData()
+                    this.dialogVisible = false
                 })
             }
         }
